Fix logcat issues: add file existence check in copyDirectory, remove debug log in ViewEditor

+++ app/src/main/java/pro/sketchware/core/EncryptedFileUtil.java
@@ -81,6 +81,10 @@ public class EncryptedFileUtil {
       errorBuilder.append(destFile.getAbsolutePath());
       throw new IOException(errorBuilder.toString());
     } 
+    if (!srcFile.exists()) {
+      Log.w("EncryptedFileUtil", "copyDirectory: source file does not exist: " + srcFile.getAbsolutePath());
+      return;
+    }
     copyFile(srcFile.getAbsolutePath(), destFile.getAbsolutePath());
   }
